Extract ReaderTracker interface to `:libs:reader`
Refactor ReaderTracker from a concrete class to an interface in
libs:reader, enabling reader files that depend on it to eventually
move to the library module.

- Create ReaderTracker interface in libs:reader with all public methods
- Move SOURCE_* constants and static utilities to the interface companion
- Move ReaderTab enum to its own file in libs:reader
- Move ReaderTypes to libs:reader (depends on SOURCE_* constants)
- Rename concrete class to ReaderTrackerImpl implementing the interface
- Add Hilt @Binds module (ReaderModule) to wire impl to interface
- Add libs:analytics dependency to libs:reader
